Python/requirements.txt
pyserial==3.5
joblib==1.5.3
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.6.1
pandas==2.2.3
